requirements.txt
streamlit
mistralai
sentence-transformers
faiss-cpu
python-dotenv
transformers
PyMuPDF
numpy
requests
